rename files and group all files under "src/pages/PedCancerDataNavPage" directory
